requirements.txt
webdriver-manager==3.8.5
selenium==4.7.2
tesseract==0.1.3
# tesseract-ocr==0.0.1
requests==2.25.1
Pillow==9.4.0
pytesseract==0.3.10